# apps/backend/docs/swagger.yaml
basePath: /
definitions:
  AssignRequest:
    properties:
      planId:
        type: string
    required:
    - planId
    type: object
  AssignRoleRequest:
    properties:
      roleId:
        type: string
      userId:
        type: string
    required:
    - roleId
    - userId
    type: object
  CreateRequest:
    properties:
      name:
        minLength: 1
        type: string
      slug:
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        minLength: 1
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    type: object
  ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  LogResponse:
    properties:
      action:
        type: string
      createdAt:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      organizationId:
        type: string
      userId:
        type: string
    type: object
...
  MembershipResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      organization:
        $ref: '#/definitions/OrgResponse'
      organizationId:
        type: string
      role:
        type: string
      userId:
        type: string
    type: object
  OrgResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  PermissionResponse:
    properties:
      action:
        type: string
      createdAt:
        type: string
      id:
        type: string
      roleId:
        type: string
    type: object
  PlanResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      limits:
        items:
          type: integer
        type: array
      name:
        type: string
    type: object
...
  RoleResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organizationId:
        type: string
      permissions:
        items:
          $ref: '#/definitions/PermissionResponse'
        type: array
    type: object
  RoleRowResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organizationId:
        type: string
    type: object
  SubscriptionResponse:
    properties:
      createdAt:
        type: string
      customLimits:
        items:
          type: integer
        type: array
      id:
        type: string
      organizationId:
        type: string
      plan:
        $ref: '#/definitions/PlanResponse'
      planId:
        type: string
      updatedAt:
        type: string
    type: object
  SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
  UpdatePermissionsRequest:
    properties:
      permissions:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: Multi-tenant B2B SaaS control-plane API (Go port of controlplane-api).
  title: Controlplane API
  version: 0.1.0
paths:
  /audit-logs:
    get:
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Max results (1-100, default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/LogResponse'
            type: array
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Query audit logs
      tags:
      - audit-logs
...
  /organizations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MembershipResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List my organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrgResponse'
        "409":
          description: SLUG_TAKEN
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an organization
      tags:
      - organizations
...
  /rbac/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Assignment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: MEMBER_NOT_FOUND / ROLE_NOT_FOUND
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign a role to a member
      tags:
      - rbac
  /rbac/roles:
    get:
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RoleResponse'
            type: array
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - rbac
    post:
      consumes:
      - application/json
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Role payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleRowResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - rbac
  /rbac/roles/{roleId}/permissions:
    put:
      consumes:
      - application/json
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: New permission set
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdatePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: ROLE_NOT_FOUND
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replace a role's permissions
      tags:
      - rbac
  /subscription:
    get:
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SubscriptionResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the organization's subscription
      tags:
      - subscription
  /subscription/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Plan payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AssignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign a subscription plan
      tags:
      - subscription
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" — the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"